refactor(builder): Split resources creation from CliBuilder to ResourcesBuilder
